requirements.txt
# requirements.txt
fpdf
PyPDF2
python-dotenv
